Improve test coverage of TorpedoManager Fixes #153

+++ src/main/groovy/net/ebdon/trk21/PhaserControl.groovy
@@ -47,7 +47,7 @@ final class PhaserControl {
 
   @SuppressWarnings('InsecureRandom')
   private float phaserVariance() {
-    2 + new Random().nextFloat()
+    (2f + new Random().nextFloat()) as float
   }
 
   private float rangeTo( final Expando target ) {
